Fix `_sanitize`
The role of `_sanitize` is to remove duplicates for a single stock data.
It should not be used for stocks data when building a consolidated
`DataFrame`. Moreover, just checking the `Date` column for duplicates is
sufficient (and necessary because the symbol column would be modified
when there is an adjustment due to corporate actions or symbol change).